Changed to int64_t. SQLite documentation states that rowid is a 64-bit signed integer key. Updated the application to match the key size returned from the database. This is intended to avoid any overflow that could occur.

+++ source/rss_lib/rss/rss_feed_mod.hpp
@@ -14,6 +14,7 @@ Author: Michael Gautier <michaelgautier.wordpress.com>
 #define michael_gautier_rss_data_read_feed_mod_h
 
 #include <string>
+#include <cstdint>
 
 namespace gautier_rss_data_read {
 	enum class
@@ -30,7 +31,7 @@ namespace gautier_rss_data_read {
 		public:
 			std::string feed_name;
 			rss_feed_mod_status status;
-			int row_id;
+			int64_t row_id;
 	};
 }
 #endif
